@@ -1,13 +1,13 @@
 import { Injectable } from '@angular/core';
-import { Http, Response } from '@angular/http';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable()
 export class DropdownService {
 
-  constructor(private http: Http) { }
+  constructor(private http: HttpClient) { }
 
   getEstadosBr() {
-    return this.http.get('assets/dados/estadosbr.json').map((res: Response) => res.json());
+    return this.http.get('assets/dados/estadosbr.json');
   }
 
 }
